major overhaul to classes and tests
