--- src/main/java/com/zipcodewilmington/scientificcalculator/Addition.java
@@ -1,13 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class Addition{
-
-    public Addition(){
-    }
-
-    public double add(double value1 , double value2 ){
-        return value1 + value2;
-    }
-}
-
-
