+++ src/main/java/com/zipcodewilmington/scientificcalculator/Console.java
@@ -18,8 +18,7 @@ public class Console {
     public static String getStringInput(String prompt) {
         Scanner scanner = new Scanner(System.in);
         println(prompt);
-        String userInput = scanner.nextLine();
-        return userInput;
+        return scanner.nextLine();
     }
 
     public static Integer getIntegerInput(String prompt) {
@@ -30,6 +29,143 @@ public class Console {
         return Double.parseDouble(getStringInput(prompt));
     }
 
+    Double add(double value1, double value2) {
+        return value1 + value2;
+    }
 
+    Double subtract(double value1, double value2) {
+        return value1 - value2;
+    }
 
+    Double multiply(double value1, double value2) {
+        return value1 * value2;
+    }
+
+    Double divide(double value1, double value2) {
+        if (value2 == 0) {
+            System.out.println("ERROR: Cannot divide by 0.");
+            return null;
+        } else {
+            return value1 / value2;
+        }
+    }
+
+    Double exponent(double value1, double value2) {
+        return Math.pow(value1, value2);
+    }
+
+    Double cosine(double value) {
+        if (value < 0) {
+            return Math.cos(-value);
+        } else {
+            return Math.cos(value);
+        }
+    }
+
+    Double factorial(Double value1) {
+        double facto = 1;
+        for (int i = 1; i <= value1; i++) {
+            facto = facto * i;
+        }
+        return facto;
+    }
+
+    Double inverse(Double value) {
+        if (value == 0) {
+            System.out.println("ERROR: Cannot have an inverse of 0.");
+            return null;
+        } else {
+            return 1 / value;
+        }
+    }
+
+    Double inverseCosine(double value) {
+        if (value <= 1 && value >= -1) {
+            return Math.acos(value);
+        } else {
+            System.out.println("ERROR: Please enter a number between -1 and 1.");
+            return null;
+        }
+    }
+
+    Double inverseLogarithm(double value) {
+        if (value <= 0) {
+            System.out.println("ERROR: Please enter a number larger than 0.");
+            return null;
+        } else {
+            return Math.pow(10, value);
+        }
+    }
+
+    Double inverseNaturalLogarithm(double value) {
+        return Math.exp(value);
+    }
+
+    Double inverseSine(double value) {
+        if (value <= 1 && value >= -1) {
+            return Math.asin(value);
+        } else {
+            System.out.println("ERROR: Please enter a number between -1 and 1.");
+            return null;
+        }
+    }
+
+    Double inverseTangent(double value) {
+        return Math.atan(value);
+    }
+
+    Double invert(double value1) {
+        return value1 * -1;
+    }
+
+    Double logarithm(double value) {
+        if (value <= 0) {
+            System.out.println("ERROR: Please enter a number larger than 0.");
+            return null;
+        } else {
+            return Math.log(value);
+        }
+    }
+
+    Double sine(double value) {
+        if (value < 0) {
+            return Math.sin(-value);
+        } else {
+            return Math.sin(value);
+        }
+    }
+
+    Double sqrt(double value) {
+        if (value <= 0) {
+            System.out.println("ERROR: Please enter a positive number.");
+            return null;
+        } else {
+            return Math.sqrt(value);
+        }
+    }
+
+    Double squared(double value1) {
+        return value1 * value1;
+    }
+
+    Double tangent(double value) {
+        if (value < 0) {
+            return Math.tan(-value);
+        } else {
+            return Math.tan(value);
+        }
+    }
+/*
+    Double naturalLogarithm(double value) {
+        if (value <= 0) {
+            System.out.println("ERROR: Please enter a number larger than 0.");
+            return null;
+        } else {
+            return (-Math.log(1 - value)) / value;
+        }
+    }*/
 }
+
+
+
+
