--- src/main/java/com/zipcodewilmington/scientificcalculator/Exponent.java
@@ -1,13 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class Exponent{
-
-    public Exponent(){
-    }
-
-    public Double exponent(double value1, double value2){
-        return  Math.pow(value1, value2);
-    }
-}
-
-
