--- src/main/java/com/zipcodewilmington/scientificcalculator/Factorial.java
@@ -1,16 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class Factorial {
-    public Factorial(){
-
-    }
-
-    public Double factorial(Double value1){
-        double facto = 1;
-        for(int i = 1; i <= value1; i++){
-                facto = facto * i;
-            }
-            return facto;
-        }
-    }
-
