+++ src/main/java/com/zipcodewilmington/scientificcalculator/InputChecker.java
@@ -1,11 +1,7 @@
 package com.zipcodewilmington.scientificcalculator;
 
 public class InputChecker {
-
-    public InputChecker(){
-    }
-
-    public static boolean isNumeric(String str) {
+    static boolean isNumeric(String str) {
         try {
             Double.parseDouble(str);
             return true;
