--- src/main/java/com/zipcodewilmington/scientificcalculator/InverseNaturalLogarithm.java
@@ -1,9 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class InverseNaturalLogarithm {
-    public InverseNaturalLogarithm(){
-    }
-    public Double inverseNaturalLogarithm(double value){
-        return Math.exp(value);
-    }
-}
