--- src/main/java/com/zipcodewilmington/scientificcalculator/InverseTangent.java
@@ -1,8 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class InverseTangent {
-    public InverseTangent(){}
-    public Double inverseTangent(double value){
-        return Math.atan(value);
-    }
-}
