--- src/main/java/com/zipcodewilmington/scientificcalculator/Invert.java
@@ -1,13 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class Invert {
-
-    public Invert() {
-    }
-
-    public Double invert(double value1) {
-        return value1 * -1;
-    }
-
-
-}
