--- src/main/java/com/zipcodewilmington/scientificcalculator/Multiplication.java
@@ -1,14 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-
-public class Multiplication{
-
-    public Multiplication(){
-    }
-
-    public Double multiply(double value1, double value2){
-        return value1 * value2;
-    }
-}
-
-
