--- src/main/java/com/zipcodewilmington/scientificcalculator/Squared.java
@@ -1,10 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class Squared{
-
-    public Squared(){
-    }
-    public Double squared(double value1) {
-        return value1 * value1;
-    }
-}
