--- src/main/java/com/zipcodewilmington/scientificcalculator/Subtraction.java
@@ -1,13 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-    public class Subtraction{
-
-    public Subtraction(){
-    }
-
-    public Double subtract(double value1, double value2){
-        return value1 - value2;
-    }
-}
-
-
