--- src/main/java/com/zipcodewilmington/scientificcalculator/Tangent.java
@@ -1,14 +1,0 @@
-package com.zipcodewilmington.scientificcalculator;
-
-public class Tangent {
-    public Tangent() {
-    }
-
-    public Double tangent(double value) {
-        if (value < 0) {
-            return Math.tan(-value);
-        } else {
-            return Math.tan(value);
-        }
-    }
-}
