--- src/test/java/com/zipcodewilmington/scientificcalculator/NaturalLogarithmTest.java
@@ -1,47 +1,0 @@
-/*
-package com.zipcodewilmington.scientificcalculator;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class NaturalLogarithmTest {
-
-    @Before
-    public void setUp() throws Exception {
-    }
-
-    @After
-    public void tearDown() throws Exception {
-    }
-
-    @Test
-    public void naturalLogarithmTest1() {
-        // given
-        NaturalLogarithm result = new NaturalLogarithm();
-        double inputValue = 1;
-        double expected = -Math.log((1-inputValue)/inputValue);
-
-        // when
-        double actual = result.naturalLogarithm(inputValue);
-
-        // then
-        assertEquals(expected, actual, .00);
-    }
-
-    @Test
-    public void naturalLogarithmTest2() {
-        // given
-        NaturalLogarithm result = new NaturalLogarithm();
-        double inputValue = 45;
-        double expected = -Math.log((1-inputValue)/inputValue);
-
-        // when
-        double actual = result.naturalLogarithm(inputValue);
-
-        // then
-        assertEquals(expected, actual, .00);
-    }
-}*/
